backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-filter==23.5
django-cors-headers==4.3.1
python-decouple==3.8